Replace Cython float format/parsing with vendored C
Remove Cython third_party wrappers (fastfloat, ryu, lz4, zstd) and
vendor the native formatting/parsing sources into the build. Wire
Draken native cast/format helpers for string<->float paths, fix AVG
result dtype and Py_ssize_t usages in pyx, and adjust tests to match
the new native behaviour. Update .gitignore for new vector objects and
add an investigation PM doc for an intermittent crash.

+++ opteryx/operators/aggregate/ungrouped_agg_engine.pyx
@@ -177,7 +177,10 @@ cdef class UngroupedAggregateEngine:
             else:
                 value = s / c
             names.append(afin.output_alias)
-            vectors.append(vector_from_sequence([value]))
+            # AVG always produces a float result (or None). Dispatch through the
+            # float64 constructor explicitly — the int64-default path errors on
+            # the float value.
+            vectors.append(vector_from_sequence([value], dtype="DOUBLE"))
 
         return Morsel.from_vectors(names, vectors)
